Move lib to src
